PHP update, Error handling fix
We got an 505 error on the server. This is because the local and the server's PHP was not on the same version so the code could not run the way as intended.
New PHP version is 8.2

After the sync I was able to see that we get the opposit result by using try-cath at our PDO, so it is taken out off that.

Also, the text of the page included a grammar error copied from the original task description.